did abstraction and upload option on desktop
